Add refactor conection db to db.js & import to services
