+++ backend/src/db.js
@@ -1,0 +1,12 @@
+const knex= require('knex');
+
+//conexión con la db
+const db = knex({
+    client: 'sqlite3',
+    connection: {
+        filename: 'boardgames.db'
+    },
+    useNullAsDefault: true
+})
+
+module.exports = db;
