+++ backend/src/service/boardgames.js
@@ -1,3 +1,6 @@
+const db = require('../db');
+
+
 const findBoardgames = (async () => {
 
 })
