+++ backend/src/service/games.js
@@ -1,3 +1,5 @@
+const db = require('../db');
+
 const findGames = (async () => {
 
 })
