+++ backend/src/service/gamesUsers.js
@@ -1,3 +1,6 @@
+const db = require('../db');
+
+
 const findUsersByGame = (async(id) => {
 
 })
