+++ backend/src/service/users.js
@@ -1,3 +1,6 @@
+const db = require('../db');
+
+
 const findUsers = (async () => {
 
 })
